chore(connectors): rename vitana_hub.ts → vitana-hub.ts (Phase 2B kebab-case)
Renames the only kebab-case-violating file in the gateway tree so the
PHASE-2B-NAMING-ENFORCEMENT workflow stops failing on every PR. The
violation was introduced by PR #744 and has blocked auto-merge for
every Dev Autopilot PR since (#785, #793, #800, #829, #838).

Filename only — the connector ID string `'vitana_hub'` stays
unchanged because it's a stable contract used in:
  - data rows (autopilot_recommendations.spec_snapshot.connector_id)
  - orb-live tool enums shown to the LLM
  - source: 'vitana_hub' on emitted events

Renaming those would break consumers and lose history.

Files:
  R services/gateway/src/connectors/media/vitana_hub.ts → vitana-hub.ts
  M services/gateway/src/connectors/index.ts (import path)

That's it. Five other prose/comment references to `vitana_hub` stay as
they're describing the same data identifier.

Follow-up (separate PR): scope the kebab-case scanner in
PHASE-2B-NAMING-ENFORCEMENT.yml to PR-changed files only via
`git diff origin/main...HEAD` so a future single violation can't
regression-block every PR again.
